WIP: refactored worker into embedding and upload workers. still need to 1) write tests for upload worker, 2) alter job_service.update_job_with_batch 3) test end to end flow
